chore: move spec back
