+++ ostd/specs/mod.rs
@@ -16,8 +16,3 @@ mod sync;
 #[allow(unused_parens)]
 #[allow(unused_braces)]
 pub mod task;
-#[allow(unused_parens)]
-#[allow(unused_braces)]
-#[allow(rustdoc::invalid_rust_codeblocks)]
-#[allow(rustdoc::invalid_html_tags)]
-pub mod util;
